Optimize bundle size
We replace the markdown package with pulldown_cmark in order to no
longer depend on regex. The regex package comes with massive code bloat.
Ideally we wouldn't ship a markdown parser at all and just do everything
on the server side, but we have some dynamic markdown content.

4.2M -> 3.5M in release builds.

With wasm-opt and brotli compression this ends up being 557K sent to the
user. Not bad, but still a bit too large for my liking.